tagsapi/requirements.txt
fastapi==0.85.1
psycopg2-binary==2.9.5
python-dotenv==0.21.0
uvicorn==0.19.0